Profile screen for company with service and backend folder update

+++ DigiCSR_backend/Controllers/MediaPostController.js
@@ -6,7 +6,10 @@ exports.GetPostById = async (req, res) => {
   const postId = req.params.id;
 
   try {
-    const post = await MediaPost.findById(postId);
+    const post = await MediaPost.findById(postId).populate(
+      "author",
+      "ngo_name"
+    );
 
     if (!post) {
       console.warn("here");
@@ -15,7 +18,11 @@ exports.GetPostById = async (req, res) => {
         .json({ success: false, message: "Post not found !!" });
     }
 
-    res.status(200).json({ success: true, postData: post });
+    const postData = { ...post._doc, author: post.author.ngo_name };
+
+    // console.log(postData);
+
+    res.status(200).json({ success: true, postData });
   } catch (err) {
     console.error(err);
     return res
@@ -26,10 +33,13 @@ exports.GetPostById = async (req, res) => {
 
 exports.GetPosts = async (req, res) => {
   try {
-       // const authorId = req.user._id;
-    // console.warn(req.user);
-    // console.log(req.userType);
-    const posts = await MediaPost.find();
+    const posts =
+      req.userType !== "ngo"
+        ? await MediaPost.find().populate("author", "ngo_name")
+        : await MediaPost.find({ author: req.user._id }).populate(
+            "author",
+            "ngo_name"
+          );
 
     if (!posts) {
       return res
@@ -37,7 +47,12 @@ exports.GetPosts = async (req, res) => {
         .json({ success: false, message: "No posts found !!" });
     }
 
-    return res.status(200).json({ success: true, postsData: posts });
+    const postsData = posts.map((post) => {
+      const postData = { ...post._doc, author: post.author.ngo_name };
+      return postData;
+    });
+
+    return res.status(200).json({ success: true, postsData });
   } catch (err) {
     console.error(err);
     res.status(500).json({ success: false, message: "Internal server error." });
@@ -46,7 +61,6 @@ exports.GetPosts = async (req, res) => {
 
 exports.CreatePost = async (req, res) => {
   try {
-
     if (req.userType !== "ngo") {
       return res
         .status(401)
@@ -57,10 +71,17 @@ exports.CreatePost = async (req, res) => {
 
     const { title, content } = req.body;
 
+    if (!title || !content) {
+      return res.status(400).json({
+        success: false,
+        message: "Title and content is require to create a post !!!",
+      });
+    }
+
     const newMediaPost = new MediaPost({
       title,
       content,
-      author: user._id
+      author: user._id,
     });
 
     const createdMediaPost = await newMediaPost.save();
@@ -76,10 +97,26 @@ exports.CreatePost = async (req, res) => {
 
 exports.UpdatePost = async (req, res) => {
   try {
+    if (req.userType !== "ngo") {
+      return res
+        .status(401)
+        .send({ success: false, message: "Not Authorized." });
+    }
+
     const postId = req.params.id;
     const { title, content, mediaUrl } = req.body;
 
-    const media = await MediaPost.findById(postId);
+    if (!title || !content) {
+      return res.status(400).json({
+        success: false,
+        message: "Title and content is require to create a post !!!",
+      });
+    }
+
+    const media = await MediaPost.findById(postId).populate(
+      "author",
+      "ngo_name"
+    );
 
     if (!media) {
       return res
@@ -103,7 +140,14 @@ exports.UpdatePost = async (req, res) => {
 
     const updatedMedia = await media.save();
 
-    return res.status(200).json(updatedMedia);
+    const postData = {
+      ...updatedMedia._doc,
+      author: updatedMedia.author.ngo_name,
+    };
+
+    // console.log(postData);
+
+    return res.status(200).json({ success: true, postData });
   } catch (err) {
     return res
       .status(500)
@@ -113,6 +157,12 @@ exports.UpdatePost = async (req, res) => {
 
 exports.DeletePost = async (req, res) => {
   try {
+    if (req.userType !== "ngo" && req.userType !== "Admin") {
+      return res
+        .status(401)
+        .send({ success: false, message: "Not Authorized." });
+    }
+
     const postId = req.params.id;
 
     const deletedMedia = await MediaPost.findByIdAndDelete(postId);
@@ -123,7 +173,9 @@ exports.DeletePost = async (req, res) => {
         .json({ success: false, message: "Media post not found" });
     }
 
-    return res.status(200).json({ message: "Media post deleted successfully" });
+    return res
+      .status(200)
+      .json({ success: true, message: "Media post deleted successfully" });
   } catch (err) {
     console.log(err);
     return res
@@ -133,6 +185,10 @@ exports.DeletePost = async (req, res) => {
 };
 
 exports.uploadFile = (req, res) => {
+  if (req.userType !== "ngo") {
+    return res.status(401).send({ success: false, message: "Not Authorized." });
+  }
+
   if (!req.fileUrl) {
     return res
       .status(400)
@@ -140,11 +196,9 @@ exports.uploadFile = (req, res) => {
   }
 
   const fileUrl = req.fileUrl;
-  return res
-    .status(200)
-    .json({
-      success: true,
-      message: "File uploaded successfully",
-      url: fileUrl,
-    });
+  return res.status(200).json({
+    success: true,
+    message: "File uploaded successfully",
+    url: fileUrl,
+  });
 };
